- cleanup - save edf/h5 in different folder structure than npz - fix problem when resample is None (float64 for mne)

+++ psg_processing/core/signal_processor.py
@@ -118,7 +118,7 @@ class SignalProcessor:
         """
         # Perform high-quality polyphase resampling
         signal_resampled = resample(
-            signal.astype(np.float64),
+            signal,
             up=output_rate,
             down=input_rate,
             method="polyphase",
